% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_test.R
\name{aoba_color}
\alias{aoba_color}
\title{Function to extract aoba johsai colors as hex codes}
\usage{
aoba_color(...)
}
\arguments{
\item{...}{character names of aoba johsai colors}
}
\value{
color hex code
}
\description{
Function to extract aoba johsai colors as hex codes
}
\examples{
aoba_color("aqua")
}
