% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_test.R
\name{aoba_pal}
\alias{aoba_pal}
\title{Color scale constructor for Aoba Johsai colors}
\usage{
aoba_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette aoba_palette}

\item{reverse}{Boolean indicating whether palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette}
}
\value{
color palette hex codes
}
\description{
Color scale constructor for Aoba Johsai colors
}
\examples{
aoba_pal("cool")
aoba_pal("warm")(5)
}
