% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Grid.R
\name{plot.Grid}
\alias{plot.Grid}
\title{Plot gridded map.}
\usage{
\method{plot}{Grid}(data, lon, lat, lonLims = c(min(lon), max(lon)),
  latLims = c(min(lat), max(lat)), lattice = NULL, breaks = NULL,
  palette = rainbow(10), window.width = 8, window.height = 6,
  pch.marker = 16, title.main = NULL, title.single = NULL,
  unit.text = NULL, cex.marker = 1, cex.main = 1, cex.single = 1,
  cex.unit = 1, cex.textcbar = 1, width.cbar = 0.1,
  disp.warnings = FALSE, export = FALSE, export.format = NULL,
  export.path = NULL, xlab = NULL, ylab = NULL, lon.points = NA,
  lat.points = NA)
}
\arguments{
\item{title.single}{(optional): in multiple plots, a character vector with the individual titles for single plots.}

\item{cex.single}{(optional): in multiple plots, numeric value giving the title expansion factor. Default: 1.}

\item{data:}{mattix or list of matrices to be plotted. When it is a list, a multiplot is generated (each element of the list is plotted in a subplot).}

\item{lon:}{array of longitudes.}

\item{lat:}{array of latitudes.}

\item{lonLims:}{2-elemtent vector with the range of longitudes to be plotted. By default all the data is plotted.}

\item{latLims:}{2-elemtent vector with the range of latitudes to be plotted. By default all the data is plotted.}

\item{lattice:}{in multiple plots, 2-element vector giving the number of rows and columns to display the multiplot.}

\item{breaks:}{vector of values defining the intervals to be used in the colorbar.}

\item{palette:}{character vector with the colors for the plot. They will be interpolated to match the number of intervals defined by breaks.}

\item{window.width:}{sets the width of the device window. Modify this argument to improve the appearance. Default:8.}

\item{window.height:}{sets the height of the device window.  Modify this argument to improve the appearance. Default:6.}

\item{title.main:}{character string with the main title.}

\item{unit.text:}{character string to be placed in the colorbar with the units.}

\item{cex.main:}{numeric value giving the title expansion factor. Default: 1.}

\item{cex.textcbar:}{numeric value giving the expansion factor of the text in the colorbar. Default:1.}

\item{cex.unit:}{numeric value giving the expansion factor of the units text. Default:1.}

\item{width.cbar:}{numeric value giving the fraction of the plot (from 0 to 1) reserved for the colorbar. Modify this argument for a better appearance. Default=0.1}

\item{disp.warnings:}{logical to set if warnings shoul be displayed.}

\item{export:}{logical. If TRUE, the figure is exported as pdf file. Default: False.}

\item{export.path:}{path (including file name) to save the plot.}

\item{export.format:}{"png" or "pdf" format.}

\item{lon.points:}{vector of longitudes for grid points to be marked in the map.}

\item{lat.points:}{vector of latitudes for grid points to be marked in the map.}
}
\description{
Plot a gridded map, either a single plot or multiplot. A common colorbar is set for a multiplot.
}
\details{
Packages classInt and maps needed.
}
\author{
Ana Casanueva (24.05.2017)
}
