% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.heat.R
\name{plot.heat}
\alias{plot.heat}
\title{Plot heat index as a function of the input variables.}
\usage{
\method{plot}{heat}(hu, ta, xlim = range(hu, na.rm = T), ylim = range(ta,
  na.rm = T), heat.index = NULL, breaks.index = NULL, title = NULL,
  xlab = "Relative Humidity", ylab = "Maximum temperature", n.bins = 500,
  add.points = TRUE, add.contours = TRUE)
}
\arguments{
\item{hu:}{vector with relative humidity data.}

\item{ta:}{vector with air temperature data.}

\item{xlim:}{2-element vector with X-axis limits for the density plot. Default: range of hu.}

\item{ylim:}{2-element vector with Y-axis limits for the density plot. Default: range of ta.}

\item{breaks.index:}{vector of breaks for the heat index values. Example: breaks.den=seq(min, max, length=1000). By default, breaks is adjusted to the minimum and maximum values of the index.}

\item{n.bins:}{number of bins for kernel density calculation. Default= 500}

\item{add.contours:}{logical. Add (default) or not contours for heat index values.}

\item{add.points:}{logical. Add (default) or not points with the eactual time series.}

\item{xlab:}{X-axis label.}

\item{ylab:}{Y-axis label.}

\item{title:}{title above the plot.}
}
\value{
Plot with the heat index values as a function of the input variables.
}
\description{
Plot heat index as a function of the input variables.
}
\details{
Needed packages: HeatStress, RColorBrewer.
}
\author{
: Ana Casanueva, 13.12.2018
}
