% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.kde2d.hist.R
\name{plot.kde2d.hist}
\alias{plot.kde2d.hist}
\title{Plot 2-dimensional kernel density plots and the histograms of the two input variables.}
\usage{
\method{plot}{kde2d.hist}(var1, var2, xlim = range(var1, na.rm = T),
  ylim = range(var2, na.rm = T), breaks.den = NULL, title = NULL,
  cex.main = 1.5, xlab = NULL, ylab = NULL, n.bins = 25,
  add.contours = TRUE)
}
\arguments{
\item{var1}{vector with data for one variable.}

\item{var2}{vector with data for another variable.}

\item{xlim}{2-element vector with X-axis limits for the density plot. Default: range of var1.}

\item{ylim}{2-element vector with Y-axis limits for the density plot. Default: range of var2.}

\item{breaks.den}{vector of breaks for the density plots. Example: breaks.den=seq(min, max, length=1000). By default, breaks.den is adjusted to the minimum and maximum kernel densities.}

\item{title}{title above the plot.}

\item{cex.main}{Relative size of the plot title. Default:1.5.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{n.bins}{number of bins for kernel density calculation. Default= 25.}

\item{add.contours}{logical. Add (default) or not contours for the kernel densities.}
}
\value{
Plot with the 2-dimensional kernel densities and historgrams for the input variables.
}
\description{
Plot 2-dimensional kernel density plots and the histograms of the two input variables.
}
\details{
The two-dimensional kernel density estimation is done with the function kde2d in package MASS.
}
\examples{
\dontrun{
# Generate data
tas <- rnorm(1000, mean=15, sd=2)
td <- rnorm(1000, mean=8, sd=1)
# Plot "-Dim Kernel density with histograms
plot.kde2d.hist(var1=td, var2=tas, xlab="Dew point temp.", ylab="Air temperature", 
n.bins=25, add.contours=FALSE, title="2D density plot", cex.main=2)
# add contours for the density values
plot.kde2d.hist(var1=td, var2=tas, xlab="Dew point temp.", ylab="Air temperature", 
n.bins=25, add.contours=TRUE, title="2D density plot with contours", cex.main=2)
}
}
\author{
Ana Casanueva, 13.12.2018
}
