% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.kde2d.hist.R
\name{plot.kde2dhist}
\alias{plot.kde2dhist}
\title{Plot 2-dimensional kernel density plots and the histograms of the two input variables.}
\usage{
\method{plot}{kde2dhist}(hu, ta, xlim = range(hu, na.rm = T),
  ylim = range(ta, na.rm = T), breaks.den = NULL, title = NULL,
  xlab = "Relative Humidity", ylab = "Maximum temperature", n.bins = 25,
  add.contours = TRUE)
}
\arguments{
\item{hu:}{vector with dew point temperature data.}

\item{ta:}{vector with air temperature data.}

\item{xlim:}{2-element vector with X-axis limits for the density plot. Default: range of hu.}

\item{ylim:}{2-element vector with Y-axis limits for the density plot. Default: range of ta.}

\item{breaks.den:}{vector of breaks for the density plots. Example: breaks.den=seq(min, max, length=1000). By default, breaks.den is adjusted to the minimum and maximum kernel densities.}

\item{n.bins:}{number of bins for kernel density calculation. Default= 25.}

\item{add.contours:}{logical. Add (default) or not contours for the kernel densities.}

\item{xlab:}{X-axis label.}

\item{ylab:}{Y-axis label.}

\item{title:}{title above the plot.}
}
\value{
Plot with the 2-dimensional kernel densities and historgrams for the input variables.
}
\description{
Plot 2-dimensional kernel density plots and the histograms of the two input variables.
}
\details{
The two-dimensional kernel density estimation is done with the function kde2d in package MASS.
}
\author{
: Ana Casanueva, 13.12.2018
}
