% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Stn.R
\name{plot.Stations}
\alias{plot.Stations}
\title{Plot map with stations.}
\usage{
\method{plot}{Stations}(data, lon, lat, xlims, ylims, lattice = NULL,
  breaks = NULL, palette = rainbow(10), window.width = 8,
  window.height = 6, title.main = NULL, title.single = NULL,
  unit.text = NULL, mark.border = TRUE, cex.marker = 1, cex.main = 1,
  cex.single = 1, cex.unit = 1, cex.textcbar = 1, width.cbar = 0.1,
  disp.warnings = FALSE, axis.lab = FALSE, export = FALSE,
  export.format = NULL, export.path = NULL)
}
\arguments{
\item{title.single}{(optional): in multiple plots, a character vector with the individual titles for single plots.}

\item{cex.single}{(optional): in multiple plots, numeric value giving the title expansion factor. Default: 1.}

\item{data:}{vector or matrix to be plotted. When it is a matrix, a multiplot is generated (each row of data is plotted in a subplot).}

\item{lon:}{vector of longitudes. It must have the same length as data (or number of columns in data).}

\item{lat:}{vector of latitudes. It must have the same length as data (or number of columns in data).}

\item{xlims:}{2-element vector defining the minimum and maximum longitud for the plotting area.}

\item{ylims:}{2-element vector defining the minimum and maximum latitude for the plotting area.}

\item{lattice:}{in multiple plots, 2-element vector giving the number of rows and columns to display the multiplot.}

\item{breaks:}{vector of values defining the intervals to be used in the colorbar.}

\item{palette:}{character vector with the colors for the plot. They will be interpolated to match the number of intervals defined by breaks.}

\item{window.width:}{sets the width of the device window. Modify this argument to improve the appearance. Default:8.}

\item{window.height:}{sets the height of the device window.  Modify this argument to improve the appearance. Default:6.}

\item{pch.marker:}{numeric value giving the symbol to use as marker. Default:16 (circle).}

\item{title.main:}{character string with the main title.}

\item{unit.text:}{character string to be placed in the colorbar with the units.}

\item{mark.border:}{logical. If black contour for the cicles is plotted. Default=TRUE.}

\item{cex.marker:}{numeric value giving the marker expansion factor. Default: 1.}

\item{cex.main:}{numeric value giving the title expansion factor. Default: 1.}

\item{cex.textcbar:}{numeric value giving the expansion factor of the text in the colorbar. Default:1.}

\item{cex.unit:}{numeric value giving the expansion factor of the units text. Default:1.}

\item{width.cbar:}{numeric value giving the fraction of the plot (from 0 to 1) reserved for the colorbar. Modify this argument for a better appearance. Default=0.1
param axis.lab: display or not lon/lat ticks and labels. Default=FALSE}

\item{disp.warnings:}{logical to set if warnings shoul be displayed.}

\item{export:}{logical. If TRUE, the figure is exported as pdf file. Default: False.}

\item{export.path:}{path (including file name) to save the plot.}

\item{export.format:}{"png" or "pdf" format.}
}
\description{
Plot a map with point stations, either a single plot or multiplot. A common colorbar is set for a multiplot.
}
\details{
Packages classInt and maps needed.
}
\author{
Ana Casanueva (17.02.2017)
}
