% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts_heatstress.R
\name{plot_ts_heatstress}
\alias{plot_ts_heatstress}
\title{Plot time series of heat stress indices}
\usage{
plot_ts_heatstress(data, heat.index = NULL, index.name = NULL, agg = NULL,
  aggFun = "mean", season = NULL, unit = unit, width = NULL,
  lang = "E", output.data = FALSE, omit.plot = FALSE,
  stn.together = FALSE, col.lines = rainbow(10))
}
\arguments{
\item{data}{List with the data and meta data as a result of \code{calc_heatstress.R}.}

\item{heat.index}{Desired heat stress index to plot (only one at a time). Possible options: 'wbt', 'wbgt.shade', 'wbgt.sun', 'hi_max', 'swbgt', 'apparentTemp', 'effectiveTemp', 'humidex', 'discomInd'.}

\item{index.name}{Name of the desired heat stress index to plot (used for plots labels). No default.}

\item{agg}{Level of temporal aggregation: daily ("D", only for hourly input data), monthly ("M"), seasonal ("S"), annual ("Y"). Default: NULL (all data is plotted using the original temporal resolution). This argument is only valid for \code{type="ts"}.}

\item{aggFun}{Function for temporal aggregation, if desired. Possible values: mean, median, min, max. Default: mean.}

\item{season}{Vector with the desired month (season) to be plotted, e.g. 8 is August, c(6,7,8) is JJA, with the operation of 'aggFun'. Default: NULL (no seasonal/monthly slice considered).}

\item{unit}{Character string with the text to be added as unit of the index.}

\item{width}{Length of time steps to be considered for the running mean. Default: NULL (no running mean). Only valid for \code{type="ts"}.}

\item{lang}{Language to be used in the plot. Possible values: "G", "F", "I", "E". Default: "E".}

\item{output.data}{Logical, return plotted data or not. Default: FALSE.}

\item{omit.plot}{Logical, set to TRUE to avoid the display of the plot. Default: FALSE.}

\item{stn.together}{Logical, set to TRUE to plot all stations with colour lines in one plot. Default: FALSE. Only available for time series plots.}

\item{col.lines}{Vector with colours for lines. Only available for time series plots.}
}
\value{
A time series plot. Plotted data if 'output.data'=TRUE.
}
\description{
Plot a heat stress object coming from \code{calc_heatstress.R}.
}
\details{
Plot time series for the desired heat stress index, level of temporal aggregation or season/month. A running mean can be added using the argument 'width'.
}
\examples{
\dontrun{
# plot daily values of WBT in Geneva and Basel with a moving window
 plot_ts_heatstress(result, sta = c("GVE", "BAS"), heat.index = c("wbt"), width=30) 
 # plot monthly maximum values of WBT in Geneva and Basel
 plot_ts_heatstress(result, sta = c("GVE", "BAS"), agg="M", aggFun="max", heat.index = c("wbt")) 
 # plot July's maximum values of WBT in Geneva and Basel
 plot_ts_heatstress(result, sta = c("GVE", "BAS"), season=7, aggFun="max", heat.index = c("wbt")) 
 # plot summer maximum values of WBT in Geneva and Basel
 plot_ts_heatstress(result, sta = c("GVE", "BAS"), season=c(6,7,8), aggFun="max", 
 heat.index = c("wbt")) 
}

}
