% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFun.barplotScen.R
\name{plotFun.barplotScen}
\alias{plotFun.barplotScen}
\title{Barplot with whiskers.}
\usage{
plotFun.barplotScen(obs, ctl, ctl.rg, p1, p1.rg, p2, p2.rg, p3, p3.rg, col,
  ylab = NULL, xlab = NULL, stn.name = NULL, breaks = seq(0,
  ceiling(p3.rg[2])), add.lines = F, lines = NULL, pch.obs = 8,
  cex.obs = 1.5)
}
\arguments{
\item{obs}{observed value.}

\item{ctl}{multi-model median/mean value in the historical period.}

\item{ctl.rg}{2-element vector with the multi.model range (e.g. 5-95th percentiles) in the historical period.}

\item{p1}{multi-model median/mean value in the short-term future period.}

\item{p1.rg}{2-element vector with the multi.model range (e.g. 5-95th percentiles) in the short-term future period.}

\item{col}{4-element vector with the colours for the four bars.}

\item{ylab}{character string for the Y-axis.}

\item{xlab}{4-element vector to label the bars.}

\item{stn.name}{character with the name of the station.}

\item{breaks}{vector of values with the Y-axis ticks.}

\item{add.lines}{should horizontal lines be added for reference? Default: FALSE.}

\item{lines}{values in the Y-axis to plot a horizontal line.}

\item{pch.obs}{type of marker for the observations. Default:8.}

\item{cex.obs}{relative size for the marker representing the observations. Default: 1.5.}

\item{p1}{multi-model median/mean value in the mid-term future period.}

\item{p1.rg}{2-element vector with the multi.model range (e.g. 5-95th percentiles) in the mid-term future period.}

\item{p1}{multi-model median/mean value in the long-term future period.}

\item{p1.rg}{2-element vector with the multi.model range (e.g. 5-95th percentiles) in the long-term future period.}
}
\value{
A bar plot consisting on 4 bars (e.g. historical and 3 future periods) and highlight the observations with a markers. Uncertainty is shown though small segments.
}
\description{
Barplot with whiskers, for instance to show mean and range of climate projections.
}
\author{
Ana Casanueva (22.03.2019)
}
