% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFun.colorbar.R
\name{plotFun.colorbar}
\alias{plotFun.colorbar}
\title{Plot the colorbar.}
\usage{
plotFun.colorbar(breaks, palette = palette, unit.text, cex.unit, cex.textcbar)
}
\arguments{
\item{breaks}{vector of values defining the intervals to be used in the colorbar.}

\item{palette}{character vector with the colors for the plot. They will be interpolated to match the number of intervals defined by breaks.}

\item{unit.text}{character string to be placed in the colorbar with the units.}

\item{cex.unit}{numeric value giving the expansion factor of the units text. Default:1.}

\item{cex.textcbar}{numeric value giving the expansion factor of the colorbar text. Default:1.}
}
\description{
Plot the colorbar.
}
\author{
Ana Casanueva (16.02.2017)
}
