% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFun.R
\docType{package}
\name{plotFun}
\alias{plotFun}
\alias{plotFun-package}
\title{plotFun}
\description{
Several functions to plot meteorological data.
}
\details{
The package \code{plotFun} provides helper functions to plot meteorological data.



The following plotting functions are implemented:
 \itemize{
 
 \item \code{plot.Stn}: Plot a map of stations.
 \item \code{plot.Grid}: Plot a map of a regular grid.
 \item \code{plot.ts}: Plot time series with or without temporal aggregation.
 \item \code{plot.kde2d.hist}: Plot a 2D Kernel density plot with the histograms of the marginal distributions.
 
 
 Additional functions:  
 \item \code{data.agg}: Function to obtain temporal aggregations (from hours to days and from days to months, seasons or years).
 \item \code{sort4plot}: Funtion to sort the data in ascending order before plot.
 }
}
