% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFun.stn.R
\name{plotFun.stn}
\alias{plotFun.stn}
\title{Plot map with stations.}
\usage{
plotFun.stn(data, lon, lat, xlims, ylims, lattice = NULL,
  breaks = NULL, palette = rainbow(10), window.width = 8,
  window.height = 6, title.main = NULL, title.single = NULL,
  unit.text = NULL, mark.border = TRUE, cex.marker = 1,
  cex.main = 1, cex.single = 1, cex.unit = 1, cex.textcbar = 1,
  width.cbar = 0.1, disp.warnings = FALSE, axis.lab = FALSE,
  export = FALSE, export.format = NULL, export.path = NULL,
  text = NULL, cex.text = 1, map.fill = FALSE, map.col = "grey89")
}
\arguments{
\item{data}{vector or matrix to be plotted. When it is a matrix, a multiplot is generated (each row of data is plotted in a subplot).}

\item{lon}{vector of longitudes. It must have the same length as data (or number of columns in data).}

\item{lat}{vector of latitudes. It must have the same length as data (or number of columns in data).}

\item{xlims}{2-element vector defining the minimum and maximum longitud for the plotting area.}

\item{ylims}{2-element vector defining the minimum and maximum latitude for the plotting area.}

\item{lattice}{in multiple plots, 2-element vector giving the number of rows and columns to display the multiplot.}

\item{breaks}{vector of values defining the intervals to be used in the colorbar.}

\item{palette}{character vector with the colors for the plot. They will be interpolated to match the number of intervals defined by breaks.}

\item{window.width}{sets the width of the device window. Modify this argument to improve the appearance. Default:8.}

\item{window.height}{sets the height of the device window.  Modify this argument to improve the appearance. Default:6.}

\item{title.main}{character string with the main title.}

\item{title.single}{(optional): in multiple plots, a character vector with the individual titles for single plots.}

\item{unit.text}{character string to be placed in the colorbar with the units.}

\item{mark.border}{logical. If black contour for the cicles is plotted. Default=TRUE.}

\item{cex.marker}{numeric value giving the marker expansion factor. Default: 1.}

\item{cex.main}{numeric value giving the title expansion factor. Default: 1.}

\item{cex.single}{(optional): in multiple plots, numeric value giving the title expansion factor. Default: 1.}

\item{cex.unit}{numeric value giving the expansion factor of the units text. Default:1.}

\item{cex.textcbar}{numeric value giving the expansion factor of the text in the colorbar. Default:1.}

\item{width.cbar}{numeric value giving the fraction of the plot (from 0 to 1) reserved for the colorbar. Modify this argument for a better appearance. Default=0.1}

\item{disp.warnings}{logical to set if warnings shoul be displayed.}

\item{axis.lab}{display or not lon/lat ticks and labels. Default=FALSE}

\item{export}{logical. If TRUE, the figure is exported as pdf file. Default: False.}

\item{export.format}{"png" or "pdf" format.}

\item{export.path}{path (including file name) to save the plot.}

\item{text}{display some text in each point. Length should be the same as lat and lon.}

\item{cex.text}{numeric value giving the text expansion factor. Default: 1.}

\item{map.fill}{logical. If map is filled with colour in `map.col'. Default: FALSE}

\item{map.col}{Colour for the map (continent). Default: grey89.}
}
\description{
Plot a map with point stations, either a single plot or multiplot. A common colorbar is set for a multiplot.
}
\details{
Packages classInt and maps needed.
}
\examples{
\dontrun{
# Generate data
tas <- cbind(rnorm(10, mean=15, sd=2), rnorm(10, mean=20, sd=1.5))
longitude = c(6,-5)
latitude = c(46, 40)
# Single plot
plotFun.stn(apply(tas,2, mean, na.rm=T), lon=longitude, lat = latitude, xlims = c(-10,30), 
ylims = c(35,65), width.cbar = 0.2, breaks = seq(10,20), title.main = "Mean Temp.", 
unit.text = "degC", cex.main=1.5, cex.marker = 2)
# Multi-plots
plotFun.stn(tas[1:2,], lon=longitude, lat = latitude, xlims = c(-10,30), ylims = c(35,65), 
width.cbar = 0.15, breaks = seq(10,20), window.height = 10, window.width = 6, 
title.single = c("Temp1","Temp2"), cex.single = 1.3, title.main = "Station plot", cex.main = 2, 
unit.text = "degC")
}
}
\author{
Ana Casanueva (17.02.2017)
}
