% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_test.R
\name{scale_color_aoba}
\alias{scale_color_aoba}
\title{Color scale constructor for Aoba Johsai colors}
\usage{
scale_color_aoba(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in aoba_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for Aoba Johsai colors
}
\examples{
 p + scale_color_aoba()
p + scale_color_aoba(discrete = FALSE, palette = "cool")
}
