% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Grid.R, R/plot.Stn.R, R/set.Breaks.R
\name{set.Breaks}
\alias{set.Breaks}
\alias{set.Breaks}
\alias{set.Breaks}
\title{Set breaks for a colorbar.}
\usage{
set.Breaks(data)

set.Breaks(data)

set.Breaks(data)
}
\arguments{
\item{data}{Vector of matrix from which the breaks are desired.}

\item{data:}{vector or matrix to be plotted}

\item{data:}{vector or matrix to be plotted}
}
\value{
breaks:  vector of values defining the intervals to be used in the colorbar.

breaks:  vector of values defining the intervals to be used in the colorbar.

A vector with breaks for plot.
}
\description{
Set breaks for a colorbar with the lowest and highest values. This is used when no breaks are given by the user.

Set breaks for a colorbar with the lowest and highest values. This is used when no breaks are given by the user.

Get breaks for plot based on the data interval.
}
\author{
Ana Casanueva (16.02.2017)

Ana Casanueva (16.02.2017)
}
