% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFun.R
\docType{package}
\name{plotFun}
\alias{plotFun}
\alias{plotFun-package}
\title{plotFun}
\description{
Several functions to plot meteorological data.
}
\details{
The package \code{plotFun} provides helper functions to plot meteorological data.



The following plotting functions are implemented:
 \itemize{
 
 \item \code{plotFun.stn}: Plot a map of stations.
 \item \code{plotFun.grid}: Plot a map of a regular grid.
 \item \code{plotFun.ts}: Plot time series with or without temporal aggregation.
 \item \code{plotFun.kde2d.hist}: Plot a 2D Kernel density plot with the histograms of the marginal distributions.
 \item \code{plotFun.heat}: Plot heat index as a function of the input variables.
 
 
 Additional functions:  
 \item \code{data_agg}: Function to obtain temporal aggregations (from hours to days and from days to months, seasons or years).
 \item \code{sort4plot}: Funtion to sort the data in ascending order before plot.
 }
}
