% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.Breaks.R
\name{set.Breaks}
\alias{set.Breaks}
\title{Get breaks for plot.}
\usage{
set.Breaks(data)
}
\arguments{
\item{data}{Vector of matrix from which the breaks are desired.}
}
\value{
A vector with breaks for plot.
}
\description{
Get breaks for plot based on the data interval.
}
